fix(review): address copilot review comments on PR #50
- use-sign-in: remove rethrow from submit(), return null on error; gate
  navigation in submitForm() on result.authenticated to prevent unhandled
  promise rejections and spurious redirects
- signin-client: type handleSubmit/onSubmit as FormEvent<HTMLFormElement>
  to eliminate unsafe cast
- get-bdd-status-snapshot: move Snapshot and related types from
  app/dashboard/bdd/types to src/server/bdd/types to decouple server data
  access from route-level UI directory; app/ re-exports for backward compat
- plan-nextNodeAppBase.prompt.md: deduplicate to single canonical copy of
  the Reusable Platform Extraction plan
- remove plan-nextNodeAppBase.prompt.md.bak backup artifact
